--- a/device/common/include/traccc/ambiguity_resolution/device/count_removable_tracks.hpp
+++ b/device/common/include/traccc/ambiguity_resolution/device/count_removable_tracks.hpp
@@ -8,6 +8,7 @@
 #pragma once
 
 // Project include(s).
+#include "traccc/definitions/primitives.hpp"
 #include "traccc/utils/pair.hpp"
 
 // VecMem include(s).
@@ -43,17 +44,17 @@ struct count_removable_tracks_payload {
     /**
      * @brief View object to the vector of measured ids per track
      */
-    vecmem::data::jagged_vector_view<const std::size_t> meas_ids_view;
+    vecmem::data::jagged_vector_view<const measurement_id_type> meas_ids_view;
 
     /**
      * @brief View object to the vector of number of measurements
      */
-    vecmem::data::vector_view<const std::size_t> n_meas_view;
+    vecmem::data::vector_view<const unsigned int> n_meas_view;
 
     /**
      * @brief View object to the unique measurement ids
      */
-    vecmem::data::vector_view<const std::size_t> unique_meas_view;
+    vecmem::data::vector_view<const measurement_id_type> unique_meas_view;
 
     /**
      * @brief View object to the number of accepted tracks per measurement
@@ -74,7 +75,7 @@ struct count_removable_tracks_payload {
     /**
      * @brief View object to measurements to remove
      */
-    vecmem::data::vector_view<std::size_t> meas_to_remove_view;
+    vecmem::data::vector_view<measurement_id_type> meas_to_remove_view;
 
     /**
      * @brief View object to thread id of measurements to remove