--- a/device/common/include/traccc/ambiguity_resolution/device/fill_vectors.hpp
+++ b/device/common/include/traccc/ambiguity_resolution/device/fill_vectors.hpp
@@ -30,12 +30,12 @@ struct fill_vectors_payload {
     /**
      * @brief View object to the vector of measured ids per track
      */
-    vecmem::data::jagged_vector_view<std::size_t> meas_ids_view;
+    vecmem::data::jagged_vector_view<measurement_id_type> meas_ids_view;
 
     /**
      * @brief View object to the measurement ids in flat vector
      */
-    vecmem::data::vector_view<std::size_t> flat_meas_ids_view;
+    vecmem::data::vector_view<measurement_id_type> flat_meas_ids_view;
 
     /**
      * @brief View object to the vector of pvalues
@@ -45,7 +45,7 @@ struct fill_vectors_payload {
     /**
      * @brief View object to the number of measurements per track
      */
-    vecmem::data::vector_view<std::size_t> n_meas_view;
+    vecmem::data::vector_view<unsigned int> n_meas_view;
 
     /**
      * @brief View object to the status of track acceptance