--- a/device/common/include/traccc/ambiguity_resolution/device/remove_tracks.hpp
+++ b/device/common/include/traccc/ambiguity_resolution/device/remove_tracks.hpp
@@ -34,22 +34,22 @@ struct remove_tracks_payload {
     /**
      * @brief View object to the vector of measured ids per track
      */
-    vecmem::data::jagged_vector_view<const std::size_t> meas_ids_view;
+    vecmem::data::jagged_vector_view<const measurement_id_type> meas_ids_view;
 
     /**
      * @brief View object to the vector of number of measurements
      */
-    vecmem::data::vector_view<const std::size_t> n_meas_view;
+    vecmem::data::vector_view<const unsigned int> n_meas_view;
 
     /**
      * @brief View object to the unique measurement ids
      */
-    vecmem::data::vector_view<const std::size_t> unique_meas_view;
+    vecmem::data::vector_view<const measurement_id_type> unique_meas_view;
 
     /**
      * @brief View object to the tracks per measurement
      */
-    vecmem::data::jagged_vector_view<const std::size_t>
+    vecmem::data::jagged_vector_view<const unsigned int>
         tracks_per_measurement_view;
 
     /**
@@ -87,7 +87,7 @@ struct remove_tracks_payload {
     /**
      * @brief View object to measurements to remove
      */
-    vecmem::data::vector_view<std::size_t> meas_to_remove_view;
+    vecmem::data::vector_view<measurement_id_type> meas_to_remove_view;
 
     /**
      * @brief View object to thread id of measurements to remove