… tests/cpu/test_ambiguity_resolution.cpp
… tests/cuda/test_ambiguity_resolution.cpp