I couldn't help testing the ambiguity resolution performance with unsigned int
Even though the measurement with `seq_example_cuda` is quite noisy, it seems the algorithm gets faster slightly

Before

```
14:28:51    CudaSeqExample                INFO      ==>Elapsed times...            File reading  (cpu)  883 ms
14:28:51    CudaSeqExample                INFO               Clusterization (cuda)  9 ms
14:28:51    CudaSeqExample                INFO         Spacepoint formation (cuda)  0 ms
14:28:51    CudaSeqExample                INFO                      Seeding (cuda)  52 ms
14:28:51    CudaSeqExample                INFO                 Track params (cuda)  0 ms
14:28:51    CudaSeqExample                INFO                Track finding (cuda)  228 ms
14:28:51    CudaSeqExample                INFO         Ambiguity resolution (cuda)  340 ms
14:28:51    CudaSeqExample                INFO                Track fitting (cuda)  65 ms
14:28:51    CudaSeqExample                INFO                           Wall time  1583 ms
```

After

```
14:22:45    CudaSeqExample                INFO      ==>Elapsed times...            File reading  (cpu)  892 ms
14:22:45    CudaSeqExample                INFO               Clusterization (cuda)  3 ms
14:22:45    CudaSeqExample                INFO         Spacepoint formation (cuda)  0 ms
14:22:45    CudaSeqExample                INFO                      Seeding (cuda)  53 ms
14:22:45    CudaSeqExample                INFO                 Track params (cuda)  0 ms
14:22:45    CudaSeqExample                INFO                Track finding (cuda)  217 ms
14:22:45    CudaSeqExample                INFO         Ambiguity resolution (cuda)  320 ms
14:22:45    CudaSeqExample                INFO                Track fitting (cuda)  66 ms
14:22:45    CudaSeqExample                INFO                           Wall time  1557 ms
```